+++ src/models/kv_cache.cpp
@@ -148,7 +148,7 @@ KV_Cache::KV_Cache(const Model& model, State& state)
 
   if (model_.device_type_ == DeviceType::CUDA && model_.use_cuda_graph_) {
     assert(past_present_share_buffer_);
-    size_t max_beam_batch_size = model_.config_->search.num_beams * model_.max_batch_size_;
+    size_t max_beam_batch_size = static_cast<size_t>(model_.config_->search.num_beams) * model_.max_batch_size_;
     sb_kv_caches_.reserve(layer_count_ * 2);
     for (int i = 0; i < layer_count_ * 2; ++i) {
       sb_kv_caches_.push_back(std::make_unique<StaticBuffer>(model_.allocator_device_, max_beam_batch_size));
