+++ src/models/position_inputs.cpp
@@ -49,7 +49,7 @@ PositionInputs::PositionInputs(const Model& model, State& state, RoamingArray<in
   attention_mask_shape_ = shape;
 
   if (model_.device_type_ == DeviceType::CUDA && model_.use_cuda_graph_) {
-    size_t max_beam_batch_size = model_.config_->search.num_beams * model_.max_batch_size_;
+    size_t max_beam_batch_size = static_cast<size_t>(model_.config_->search.num_beams) * model_.max_batch_size_;
     sb_position_ids_ = std::make_unique<StaticBuffer>(model_.allocator_device_, max_beam_batch_size);
     sb_seqlens_k_ = std::make_unique<StaticBuffer>(model_.allocator_device_, max_beam_batch_size);
   }
